refactor security
